requirements.txt
black==23.9.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
coverage==7.3.1
exceptiongroup==1.1.3
flake8==6.1.0
Flask==2.3.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
psycopg2-binary==2.9.8
pycodestyle==2.11.0
pyflakes==3.1.0
pytest==7.4.2
pytest-cov==4.1.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.21
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.4
Werkzeug==2.3.7
